evaluador economico

+++ app/src/main/java/com/example/tecnisis/domain/models/Solicitud.kt
@@ -7,6 +7,6 @@ data class Solicitud (
     val id_evaluador_artistico: Int,
     val aprobadaEvaluadorArtistico: Boolean,
     val aprobadaEValuadorEconomico: Boolean,
-    val porcentaje_ganancia: Double,
-    val precio_venta: Double
+    val porcentaje_ganancia: Int,
+    val precio_venta: Int
 )
